move all meta related methods from RODSConn to RC_meta trait
